Essentials: Unity Physics implemented

+++ Runtime/Modules/ViewsModule/Providers/Unity/GameObject/UnityGameObjectProvider.cs
@@ -268,7 +268,7 @@ namespace ME.ECS.Views.Providers {
         /// <summary>
         /// Time to get ready this instance to be used again after it has been despawned.
         /// </summary>
-        public virtual float cacheTimeout => this.defaultParameters.useCacheTimeout == true ? this.defaultParameters.cacheTimeout : Worlds.currentWorld.GetTimeFromTick(Worlds.currentWorld.GetModule<ME.ECS.StatesHistory.IStatesHistoryModuleBase>().GetCacheSize());
+        public virtual float cacheTimeout => this.defaultParameters.useCacheTimeout == true ? this.defaultParameters.cacheTimeout : (float)Worlds.currentWorld.GetTimeFromTick(Worlds.currentWorld.GetModule<ME.ECS.StatesHistory.IStatesHistoryModuleBase>().GetCacheSize());
         /// <summary>
         /// Time to despawn view before it has been pooled.
         /// </summary>
